pyproject.toml
[project]
name = "ratiopath"
version = "1.0.0"
description = ""
authors = [
    { name = "Matěj Pekár", email = "matejpekar@mail.muni.cz" },
    { name = "Jakub Pekár", email = "jakubpekar@mail.muni.cz" },
    { name = "Adam Kukučka", email = "adamkukucka@mail.muni.cz" },
]
readme = "README.md"
license = { file = "LICENSE" }
requires-python = ">=3.11"
dependencies = [
    "albumentations>=2.0.8",
    "imagecodecs>=2025.8.2",
    "numpy>=2.2.2",
    "ome-types>=0.6.1",
    "openslide-python>=1.4.1",
    "pandas>=2.2.3",
    "pandas-stubs>=2.2.3.250308",
    "pillow>=11.1.0",
    "pyvips>=2.2.3",
    "ray[data]>=2.42.1",
    "scikit-image>=0.25.2",
    "tifffile>=2024.5.22",
    "shapely>=2.0.0",
    "torch>=2.6.0",
    "zarr>=3.1.1",
    "geopandas>=1.1.1",
]

[project.optional-dependencies]
docs = ["mkdocs-material>=9.6.18", "mkdocstrings[python]>=0.30.0"]
tests = ["openslide-bin>=4.0.0.8", "pytest>=8.4.1"]


[tool.uv]
dev-dependencies = ["mypy", "ruff"]

[project.urls]
"Homepage" = "https://github.com/RationAI/ratiopath"
